+++ test/test_sample_order.py
@@ -11,15 +11,11 @@ class OrderExample(bn.ParametrizedSweep):
     # RESULTS
     call_index = bn.ResultVar()
 
-    def __call__(self, **kwargs):
-        self.update_params_from_kwargs(**kwargs)
-
+    def benchmark(self):
         # Maintain a per-instance counter to reflect traversal order
         idx = getattr(self, "_call_counter", 0)
         self.call_index = idx
         setattr(self, "_call_counter", idx + 1)
-
-        return super().__call__()
 
 
 class TestSampleOrder(unittest.TestCase):